+++ labbook_BE/app/services/PdfRest.py
@@ -13,16 +13,25 @@ class PdfBarcode(Resource):
     log = logging.getLogger('log_services')
 
     def get(self, num):
-        try:
-            ret = Pdf.getPdfBarcode(num)
+        ret = Pdf.getPdfBarcode(num)
 
-            if not ret:
-                self.log.error(Logs.fileline() + ' : PdfBarcode failed num=%s', str(num))
-                return compose_ret('', Constants.cst_content_type_json, 500)
-
-        except Exception as err:
-            self.log.error(Logs.fileline() + ' : PdfBarcode failed, err=%s , num=%s', err, str(num))
+        if not ret:
+            self.log.error(Logs.fileline() + ' : PdfBarcode failed num=%s', str(num))
             return compose_ret('', Constants.cst_content_type_json, 500)
 
         self.log.info(Logs.fileline() + ' : TRACE PdfBarcode')
+        return compose_ret('', Constants.cst_content_type_json)
+
+
+class PdfBill(Resource):
+    log = logging.getLogger('log_services')
+
+    def get(self, id_rec):
+        ret = Pdf.getPdfBill(id_rec)
+
+        if not ret:
+            self.log.error(Logs.fileline() + ' : PdfBill failed num=%s', str(num))
+            return compose_ret('', Constants.cst_content_type_json, 500)
+
+        self.log.info(Logs.fileline() + ' : TRACE PdfBill')
         return compose_ret('', Constants.cst_content_type_json)